some renaming and shuffling
