+++ pkg/expr/classic/classic.go
@@ -4,24 +4,21 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"math"
-	"sort"
 
-	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/grafana/grafana/pkg/expr/mathexp"
 )
 
-// ClassicConditionsCmd is command for the classic conditions
+// ConditionsCmd is command for the classic conditions
 // expression operation.
-type ClassicConditionsCmd struct {
-	Conditions []ClassicCondition
+type ConditionsCmd struct {
+	Conditions []condition
 	refID      string
 }
 
 // classicConditionJSON is the JSON model for a single condition.
 // It is based on services/alerting/conditions/query.go's newQueryCondition().
 type classicConditionJSON struct {
-	Evaluator classicConditionEvalJSON `json:"evaluator"`
+	Evaluator conditionEvalJSON `json:"evaluator"`
 
 	Operator struct {
 		Type string `json:"type"`
@@ -37,17 +34,17 @@ type classicConditionJSON struct {
 	}
 }
 
-type classicConditionEvalJSON struct {
+type conditionEvalJSON struct {
 	Params []float64 `json:"params"`
 	Type   string    `json:"type"` // e.g. "gt"
 
 }
 
-// ClassicCondition is a single condition within the ClassicConditionsCommand.
-type ClassicCondition struct {
+// condition is a single condition within the ConditionsCmd.
+type condition struct {
 	QueryRefID string
 	Reducer    classicReducer
-	Evaluator  classicEvaluator
+	Evaluator  evaluator
 	Operator   string
 }
 
@@ -55,7 +52,7 @@ type classicReducer string
 
 // NeedsVars returns the variable names (refIds) that are dependencies
 // to execute the command and allows the command to fulfill the Command interface.
-func (ccc *ClassicConditionsCmd) NeedsVars() []string {
+func (ccc *ConditionsCmd) NeedsVars() []string {
 	vars := []string{}
 	for _, c := range ccc.Conditions {
 		vars = append(vars, c.QueryRefID)
@@ -65,7 +62,7 @@ func (ccc *ClassicConditionsCmd) NeedsVars() []string {
 
 // Execute runs the command and returns the results or an error if the command
 // failed to execute.
-func (ccc *ClassicConditionsCmd) Execute(ctx context.Context, vars mathexp.Vars) (mathexp.Results, error) {
+func (ccc *ConditionsCmd) Execute(ctx context.Context, vars mathexp.Vars) (mathexp.Results, error) {
 	firing := true
 	newRes := mathexp.Results{}
 
@@ -108,8 +105,8 @@ func (ccc *ClassicConditionsCmd) Execute(ctx context.Context, vars mathexp.Vars)
 	return newRes, nil
 }
 
-// UnmarshalClassicConditionsCmd creates a new ClassicConditionsCmd.
-func UnmarshalClassicConditionsCmd(rawQuery map[string]interface{}, refID string) (*ClassicConditionsCmd, error) {
+// UnmarshalConditionsCmd creates a new ConditionsCmd.
+func UnmarshalConditionsCmd(rawQuery map[string]interface{}, refID string) (*ConditionsCmd, error) {
 	jsonFromM, err := json.Marshal(rawQuery["conditions"])
 	if err != nil {
 		return nil, fmt.Errorf("failed to remarshal classic condition body: %w", err)
@@ -119,12 +116,12 @@ func UnmarshalClassicConditionsCmd(rawQuery map[string]interface{}, refID string
 		return nil, fmt.Errorf("failed to unmarshal remarshaled classic condition body: %w", err)
 	}
 
-	c := &ClassicConditionsCmd{
+	c := &ConditionsCmd{
 		refID: refID,
 	}
 
 	for i, cj := range ccj {
-		cond := ClassicCondition{}
+		cond := condition{}
 
 		if len(cj.Query.Params) == 0 || cj.Query.Params[0] == "" {
 			return nil, fmt.Errorf("classic condition %v is missing the query refID argument", i+1)
@@ -137,7 +134,7 @@ func UnmarshalClassicConditionsCmd(rawQuery map[string]interface{}, refID string
 			return nil, fmt.Errorf("reducer '%v' in condition %v is not a valid reducer", cond.Reducer, i+1)
 		}
 
-		cond.Evaluator, err = newClassicAlertEvaluator(cj.Evaluator)
+		cond.Evaluator, err = newAlertEvaluator(cj.Evaluator)
 		if err != nil {
 			return nil, err
 		}
@@ -158,39 +155,39 @@ func (cr classicReducer) ValidReduceFunc() bool {
 	return false
 }
 
-type classicEvaluator interface {
+type evaluator interface {
 	Eval(mathexp.Number) bool
 }
 
-type classicNoValueEvaluator struct{}
+type noValueEvaluator struct{}
 
-type classicThresholdEvaluator struct {
+type thresholdEvaluator struct {
 	Type      string
 	Threshold float64
 }
 
-type classicRangedEvaluator struct {
+type rangedEvaluator struct {
 	Type  string
 	Lower float64
 	Upper float64
 }
 
-// newClassicAlertEvaluator is a factory function for returning
-// an `AlertEvaluator` depending on evalution operator.
-func newClassicAlertEvaluator(model classicConditionEvalJSON) (classicEvaluator, error) {
+// newAlertEvaluator is a factory function for returning
+// an AlertEvaluator depending on evaluation operator.
+func newAlertEvaluator(model conditionEvalJSON) (evaluator, error) {
 	switch model.Type {
 	case "gt", "lt":
-		return newClassicThresholdEvaluator(model)
+		return newThresholdEvaluator(model)
 	case "within_range", "outside_range":
-		return newClassicRangedEvaluator(model)
+		return newRangedEvaluator(model)
 	case "no_value":
-		return &classicNoValueEvaluator{}, nil
+		return &noValueEvaluator{}, nil
 	}
 
 	return nil, fmt.Errorf("evaluator invalid evaluator type: %s", model.Type)
 }
 
-func (e *classicThresholdEvaluator) Eval(reducedValue mathexp.Number) bool {
+func (e *thresholdEvaluator) Eval(reducedValue mathexp.Number) bool {
 	fv := reducedValue.GetFloat64Value()
 	if fv == nil {
 		return false
@@ -206,34 +203,34 @@ func (e *classicThresholdEvaluator) Eval(reducedValue mathexp.Number) bool {
 	return false
 }
 
-func newClassicThresholdEvaluator(model classicConditionEvalJSON) (*classicThresholdEvaluator, error) {
+func newThresholdEvaluator(model conditionEvalJSON) (*thresholdEvaluator, error) {
 	if len(model.Params) == 0 {
 		return nil, fmt.Errorf("evaluator '%v' is missing the threshold parameter", model.Type)
 	}
 
-	return &classicThresholdEvaluator{
+	return &thresholdEvaluator{
 		Type:      model.Type,
 		Threshold: model.Params[0],
 	}, nil
 }
 
-func (e *classicNoValueEvaluator) Eval(reducedValue mathexp.Number) bool {
+func (e *noValueEvaluator) Eval(reducedValue mathexp.Number) bool {
 	return !(reducedValue.GetFloat64Value() == nil)
 }
 
-func newClassicRangedEvaluator(model classicConditionEvalJSON) (*classicRangedEvaluator, error) {
+func newRangedEvaluator(model conditionEvalJSON) (*rangedEvaluator, error) {
 	if len(model.Params) != 2 {
 		return nil, fmt.Errorf("ranged evaluator requires 2 parameters")
 	}
 
-	return &classicRangedEvaluator{
+	return &rangedEvaluator{
 		Type:  model.Type,
 		Lower: model.Params[0],
 		Upper: model.Params[1],
 	}, nil
 }
 
-func (e *classicRangedEvaluator) Eval(reducedValue mathexp.Number) bool {
+func (e *rangedEvaluator) Eval(reducedValue mathexp.Number) bool {
 	fv := reducedValue.GetFloat64Value()
 	if fv == nil {
 		return false
@@ -247,181 +244,4 @@ func (e *classicRangedEvaluator) Eval(reducedValue mathexp.Number) bool {
 	}
 
 	return false
-}
-
-func nilOrNaN(f *float64) bool {
-	return f == nil || math.IsNaN(*f)
-}
-
-//nolint: gocyclo
-func (cr classicReducer) Reduce(series mathexp.Series) mathexp.Number {
-	num := mathexp.NewNumber("", nil)
-	num.SetValue(nil)
-
-	if series.Len() == 0 {
-		return num
-	}
-
-	value := float64(0)
-	allNull := true
-
-	vF := series.Frame.Fields[series.ValueIdx]
-
-	switch cr {
-	case "avg":
-		validPointsCount := 0
-		for i := 0; i < vF.Len(); i++ {
-			if f, ok := vF.At(i).(*float64); ok {
-				if nilOrNaN(f) {
-					continue
-				}
-				value += *f
-				validPointsCount++
-				allNull = false
-			}
-		}
-		if validPointsCount > 0 {
-			value /= float64(validPointsCount)
-		}
-	case "sum":
-		for i := 0; i < vF.Len(); i++ {
-			if f, ok := vF.At(i).(*float64); ok {
-				if nilOrNaN(f) {
-					continue
-				}
-				value += *f
-			}
-		}
-	case "min":
-		value = math.MaxFloat64
-		for i := 0; i < vF.Len(); i++ {
-			if f, ok := vF.At(i).(*float64); ok {
-				if nilOrNaN(f) {
-					continue
-				}
-				if value > *f {
-					value = *f
-				}
-			}
-		}
-	case "max":
-		value = -math.MaxFloat64
-		for i := 0; i < vF.Len(); i++ {
-			if f, ok := vF.At(i).(*float64); ok {
-				if nilOrNaN(f) {
-					continue
-				}
-				if value < *f {
-					value = *f
-				}
-			}
-		}
-	case "count":
-		value = float64(vF.Len())
-		allNull = false
-	case "last":
-		for i := vF.Len() - 1; i >= 0; i-- {
-			if f, ok := vF.At(i).(*float64); ok {
-				if !nilOrNaN(f) {
-					value = *f
-					allNull = false
-					break
-				}
-			}
-		}
-	case "median":
-		var values []float64
-		for i := 0; i < vF.Len(); i++ {
-			if f, ok := vF.At(i).(*float64); ok {
-				if nilOrNaN(f) {
-					continue
-				}
-				allNull = false
-				values = append(values, *f)
-			}
-		}
-		if len(values) >= 1 {
-			sort.Float64s(values)
-			length := len(values)
-			if length%2 == 1 {
-				value = values[(length-1)/2]
-			} else {
-				value = (values[(length/2)-1] + values[length/2]) / 2
-			}
-		}
-	case "diff":
-		allNull, value = calculateDiff(vF, allNull, value, diff)
-	case "diff_abs":
-		allNull, value = calculateDiff(vF, allNull, value, diffAbs)
-	case "percent_diff":
-		allNull, value = calculateDiff(vF, allNull, value, percentDiff)
-	case "percent_diff_abs":
-		allNull, value = calculateDiff(vF, allNull, value, percentDiffAbs)
-	case "count_non_null":
-		for i := 0; i < vF.Len(); i++ {
-			if f, ok := vF.At(i).(*float64); ok {
-				if nilOrNaN(f) {
-					continue
-				}
-				value++
-			}
-		}
-
-		if value > 0 {
-			allNull = false
-		}
-	}
-
-	if allNull {
-		return num
-	}
-
-	num.SetValue(&value)
-	return num
-}
-
-func calculateDiff(vF *data.Field, allNull bool, value float64, fn func(float64, float64) float64) (bool, float64) {
-	var (
-		first float64
-		i     int
-	)
-	// get the newest point
-	for i = vF.Len() - 1; i >= 0; i-- {
-		if f, ok := vF.At(i).(*float64); ok {
-			if !nilOrNaN(f) {
-				first = *f
-				allNull = false
-				break
-			}
-		}
-	}
-	if i >= 1 {
-		// get the oldest point
-		for i := 0; i < vF.Len(); i++ {
-			if f, ok := vF.At(i).(*float64); ok {
-				if !nilOrNaN(f) {
-					value = fn(first, *f)
-					allNull = false
-					break
-				}
-			}
-		}
-	}
-	return allNull, value
-}
-
-var diff = func(newest, oldest float64) float64 {
-	return newest - oldest
-}
-
-var diffAbs = func(newest, oldest float64) float64 {
-	return math.Abs(newest - oldest)
-}
-
-var percentDiff = func(newest, oldest float64) float64 {
-	return (newest - oldest) / math.Abs(oldest) * 100
-}
-
-var percentDiffAbs = func(newest, oldest float64) float64 {
-	return math.Abs((newest - oldest) / oldest * 100)
 }
